wikis for water scarcity/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
passlib==1.7.4
email-validator==1.1.3
Flask-WTF==1.0.0
python-dotenv==0.19.0
pandas==2.2.3
openpyxl==3.1.5